Add all changes.
